Move policy files
