+++ msdm/algorithms/rmax.py
@@ -9,7 +9,7 @@ import numpy as np
 from msdm.core.distributions import DictDistribution
 from msdm.core.algorithmclasses import Learns, Result
 from msdm.core.problemclasses.mdp import TabularMarkovDecisionProcess
-from msdm.core.problemclasses.mdp.policy.policy import FunctionalPolicy
+from msdm.core.problemclasses.mdp.policy import FunctionalPolicy
 from msdm.core.utils.funcutils import cached_property
 from abc import abstractmethod, ABC
