+++ msdm/core/problemclasses/mdp/__init__.py
@@ -3,3 +3,4 @@ from msdm.core.problemclasses.mdp.deterministic_shortest_path import *
 from msdm.core.problemclasses.mdp.tabularmdp import *
 from msdm.core.problemclasses.mdp.quickmdp import QuickMDP, QuickTabularMDP
 from msdm.core.problemclasses.mdp.policy import *
+from msdm.core.problemclasses.mdp.tabularpolicy import *
