--- msdm/core/problemclasses/mdp/policy/__init__.py
@@ -1,2 +1,0 @@
-from msdm.core.problemclasses.mdp.policy.tabularpolicy import TabularPolicy
-from msdm.core.problemclasses.mdp.policy.policy import Policy
